feat : domain -> entity 패키지 전략 수정 및 Provider Enum 으로 인한 코드 수정
