+++ src/main/java/com/oseak/myFestaBackend/repository/MemberPasswordRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.oseak.myFestaBackend.domain.MemberPassword;
+import com.oseak.myFestaBackend.entity.MemberPassword;
 
 public interface MemberPasswordRepository extends JpaRepository<MemberPassword, Long> {
