+++ src/main/java/com/oseak/myFestaBackend/repository/MemberRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.oseak.myFestaBackend.domain.Member;
+import com.oseak.myFestaBackend.entity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
 	Optional<Member> findByEmail(String email);
